Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 int main() {
-    int x,y; char op; scanf("%d %d %c", &x,&y,&op); 
+    float x,y; char op; scanf("%f %f %c", &x,&y,&op); 
     if(op=='+'){
-        printf("%d", x+y);
+        printf("%f", x+y);
     }
     else if(op=='-'){
-        printf("%d", x-y);
+        printf("%f", x-y);
     }
     else if(op=='*'){
-        printf("%d", x*y);
+        printf("%f", x*y);
     }
     else{
-        printf("%d", x/y);
+        printf("%f", x/y);
     }
 
     return 0;
